Add GetExerciseProgress to Analyitcs and GetNetExerciseValue to utils

+++ Models/AnalyticsItem.cs
@@ -3,6 +3,6 @@ namespace RobicServer.Models
     public class AnalyticsItem
     {
         public string label { get; set; }
-        public int count { get; set; }
+        public double count { get; set; }
     }
 }
